release

+++ app/infrastructure/adapters/out/companies_orm_repository.py
@@ -1,3 +1,4 @@
+import dataclasses
 from sqlalchemy.orm import Session
 from app.domain.entities.company_model import Company
 from app.domain.ports.out.company_repository import CompanyRepository
@@ -17,13 +18,14 @@ class CompanyORMRepository(CompanyRepository):
                 address=company.address,
                 phone=company.phone,
                 email=company.email,
-                is_active=company.is_active
+                is_active=company.is_active,
+                logo=company.logo
             )
             for company in companies
         ]
 
     def create_company(self, company: Company) -> Company:
-        company_orm = CompanyORM(**company.dict())
+        company_orm = CompanyORM(**dataclasses.asdict(company))
         self.session.add(company_orm)
         self.session.commit()
         self.session.refresh(company_orm)
@@ -34,7 +36,8 @@ class CompanyORMRepository(CompanyRepository):
             address=company_orm.address,
             phone=company_orm.phone,
             email=company_orm.email,
-            is_active=company_orm.is_active
+            is_active=company_orm.is_active,
+            logo=company_orm.logo
         )
 
     def update_company(self, id_company: str, company_data: dict) -> Company:
@@ -53,16 +56,20 @@ class CompanyORMRepository(CompanyRepository):
             address=company_orm.address,
             phone=company_orm.phone,
             email=company_orm.email,
-            is_active=company_orm.is_active
+            is_active=company_orm.is_active,
+            logo=company_orm.logo
         )
 
     def delete_company(self, id_company: str) -> None:
         company_orm = self.session.query(CompanyORM).filter(CompanyORM.id_company == id_company).first()
-        self.session.delete(company_orm)
-        self.session.commit()
+        if company_orm:
+            self.session.delete(company_orm)
+            self.session.commit()
 
     def get_company_by_id(self, id_company: str) -> Company:
         company_orm = self.session.query(CompanyORM).filter(CompanyORM.id_company == id_company).first()
+        if company_orm is None:
+            return None
         return Company(
             id_company=company_orm.id_company,
             name=company_orm.name,
@@ -70,11 +77,14 @@ class CompanyORMRepository(CompanyRepository):
             address=company_orm.address,
             phone=company_orm.phone,
             email=company_orm.email,
-            is_active=company_orm.is_active
+            is_active=company_orm.is_active,
+            logo=company_orm.logo
         )
 
     def get_company_by_nit(self, nit: str) -> Company:
         company_orm = self.session.query(CompanyORM).filter(CompanyORM.nit == nit).first()
+        if company_orm is None:
+            return None
         return Company(
             id_company=company_orm.id_company,
             name=company_orm.name,
